Further implementation and refining of CRUD and database
Implemented many of the elements needed to work towards proving the core functionality.

All fields in the database are now using int where possible as it simplifies things greatly and can always be changed again later.

+++ src/main/java/com/storeii/nciproject/model/AddressController.java
@@ -33,7 +33,7 @@ public class AddressController {
     private AddressRepository addressRepository;
     
     
-    // Adding new
+    // Add new
     @PostMapping(path="/addAddress")
     public String addAddress (
         @RequestParam String addressLine1,
@@ -65,7 +65,7 @@ public class AddressController {
     
     // find by id
     @GetMapping(path="/getAddressesById")
-    public Address getById(Long l) {
+    public Address getById(int l) {
       return addressRepository.getById(l);
     }
 }
